+++ test/unit/operation_test.js
@@ -2133,7 +2133,7 @@ describe('Operation', function() {
     it('creates a revokeTrustlineSponsorship for a liquidity pool', function() {
       const account =
         'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7';
-      const asset = new StellarBase.TrustLineAsset(
+      const asset = new StellarBase.LiquidityPoolId(
         'dd7b1ab831c273310ddbec6f97870aa83c2fbd78ce22aded37ecbf4f3380fac7'
       );
       const op = StellarBase.Operation.revokeTrustlineSponsorship({
@@ -2162,7 +2162,7 @@ describe('Operation', function() {
         StellarBase.Operation.revokeTrustlineSponsorship({
           account: 'GDGU5OAPHNPU5UCLE5RDJHG7PXZFQYWKCFOEXSXNMR6KRQRI5T6XXCD7'
         })
-      ).to.throw(/asset must be an Asset or TrustLineAsset/);
+      ).to.throw(/asset must be an Asset or LiquidityPoolId/);
     });
   });